refactor server functions
