+++ src/routes/blog/[slug]/+page.server.ts
@@ -1,9 +1,9 @@
 import type { PageServerLoad } from './$types';
 
 import { error } from '@sveltejs/kit';
-import { getPosts } from '../posts';
+import { getPosts } from '$lib/server/posts';
 
-export const load = (({ params }) => {
+export const load: PageServerLoad = async ({ params }) => {
     const post = getPosts().find((post) => post.slug === params.slug);
 
     if (!post) throw error(404);
@@ -14,4 +14,4 @@ export const load = (({ params }) => {
         createdOn: post.createdOn,
         updatedOn: post.updatedOn
     };
-}) satisfies PageServerLoad;
+};
